Allowing specifying a ACM Cert for the app (#58)
* Allowing specifying a ACM Cert for the app

This allows specifying a ACM Certificate for running the load balancer app. This allow the management of certificate from the outside.

The app uses default "internal" cert and keys when the ACM certificate is present.

The default certificate was generated against the `vpc.example.com` domain. The domain is needed, otherwise the certificate is "invalid" for use with an ALB. I also added a comment in the parameter description regarding the requirements for the certificate

Also added a load balancer for the non-HA mode, to unify our certificate handling.

* Addressed Christian's comments.