Using the VPC DNS (#46)
By default, ubuntu uses a local DNS stub. This doesn't work for Moderna, as it will not forward DNS request for the `.local` domain.

This PR changes how DNS resolution works for our App instances. By default, the configuration *"file lists the 127.0.0.53 DNS stub (see above) as the only DNS server. It also contains a list of search domains that are in use by systemd-resolved."* As `.local` is reserved for mDNS request (think local home network, where the DNS resolution is broadcast to all nodes).

After the change, *"local clients that bypass any local DNS API will also bypass systemd-resolved and will talk directly to the known DNS servers."* This is the behaviour that we want, and the default one we use for containers.