Resizing the docker disk partition on boot (#196)
While the underlying block device is at the correct size (60 GB for the app instances), the file system still thinks it has 30 GB.

This has caused multiple disk alarms to be triggered. This fixes this by resizing the docker partition `/dev/nvme1n1` during the initial setup.

The command will be successful (i.e. return 0) even if there's nothing to do, so it's safe to always run it.