Reduce user-data size by moving db-init.sh into the intermediate bucket (#202)
We are currently unable to add the Microsoft Lists secret to eMoney because adding a new secret makes the user-data longer than the allowed limit, and we get the following error when fetching the terraform plan:
```
Error: expected length of user_data to be in the range (0 - 16384)
```

This change moves the db-init.sh script out of the user-data and uploads it to the customer's intermediate bucket instead. This should make the user-data small enough that deployments can specify a few more secrets than what they are currently able to today.